+++ src/App.jsx
@@ -1,5 +1,4 @@
-import React from 'react';
-import Router from './shared/router';
+import Router from './shared/Router';
 
 const App = () => {
   return <Router />;
