+++ src/components/login/LoginModal.jsx
@@ -1,0 +1,5 @@
+const LoginModal = () => {
+  return <p>123123</p>;
+};
+
+export default LoginModal;
