+++ src/page/SignUp.jsx
@@ -1,0 +1,5 @@
+const SignUp = () => {
+  return <div>SingUp</div>;
+};
+
+export default SignUp;
