--- src/page/SingUp.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const SingUp = () => {
-  return <div>SingUp</div>;
-};
-
-export default SingUp;
