+++ src/shared/Router.jsx
@@ -1,0 +1,22 @@
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import Home from '../page/Home';
+import Login from '../page/Login';
+import MainPage from '../page/MainPage';
+import MyPage from '../page/MyPage';
+import SignUp from '../page/SignUp';
+
+const Router = () => {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="myPage" element={<MyPage />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/singup" element={<SignUp />} />
+        <Route path="mainPage" element={<MainPage />} />
+      </Routes>
+    </BrowserRouter>
+  );
+};
+
+export default Router;
